created build_models.py to build models and replace that code
